Refactor TUI, Split code and Add styled messge block
